// src/order/models/enums/order-status.enum.ts
export enum OrderStatus {
	PENDING_WAREHOUSE_RESPONSE = "PENDING_WAREHOUSE_RESPONSE",
	PROCESSING_BY_WAREHOUSE = "PROCESSING_BY_WAREHOUSE",
	WAREHOUSE_NOT_READY = "WAREHOUSE_NOT_READY",
	SHIPPED = "SHIPPED",
	DELIVERED = "DELIVERED",
	RETURN_INITIATED = "RETURN_INITIATED",
	RETURNED = "RETURNED",
	REFUNDED = "REFUNDED",
}